cleaning up modal data flow
