+++ app/src/recoil/selectors.ts
@@ -661,82 +661,6 @@ export const scalarTypes = selectorFamily({
   },
 });
 
-const COUNT_CLS = "Count";
-
-const catchLabelCount = (
-  names: string[],
-  prefix: string,
-  cur: { name: string; _CLS: string; result: number },
-  acc: { [key: string]: number }
-): void => {
-  if (
-    cur.name &&
-    names.includes(cur.name.slice(prefix.length).split(".")[0]) &&
-    cur._CLS === COUNT_CLS
-  ) {
-    acc[cur.name.slice(prefix.length).split(".")[0]] = cur.result;
-  }
-};
-
-interface Counts {
-  [key: string]: number | null;
-}
-
-export const labelSampleCounts = selectorFamily<Counts | null, string>({
-  key: "labelSampleCounts",
-  get: (dimension) => ({ get }) => {
-    const names = get(labelNames(dimension)).concat(
-      get(scalarNames(dimension))
-    );
-    const prefix = dimension === "sample" ? "" : "frames.";
-    const stats = get(datasetStats);
-    if (stats === null) {
-      return null;
-    }
-    return stats.reduce((acc, cur) => {
-      catchLabelCount(names, prefix, cur, acc);
-      return acc;
-    }, {});
-  },
-});
-
-export const filteredLabelSampleCounts = selectorFamily<Counts | null, string>({
-  key: "filteredLabelSampleCounts",
-  get: (dimension) => ({ get }) => {
-    const names = get(labelNames(dimension)).concat(
-      get(scalarNames(dimension))
-    );
-    const prefix = dimension === "sample" ? "" : "frames.";
-    const stats = get(extendedDatasetStats);
-    if (stats === null) {
-      return null;
-    }
-    return stats.reduce((acc, cur) => {
-      catchLabelCount(names, prefix, cur, acc);
-      return acc;
-    }, {});
-  },
-});
-
-export const labelSampleModalCounts = selectorFamily<Counts | null, string>({
-  key: "labelSampleModalCounts",
-  get: (dimension) => ({ get }) => {
-    // todo
-    return {};
-  },
-});
-
-export const filteredLabelSampleModalCounts = selectorFamily<
-  Counts | null,
-  string
->({
-  key: "filteredLabelSampleModalCounts",
-  get: (dimension) => ({ get }) => {
-    // todo
-    return {};
-  },
-});
-
 export const labelTuples = selectorFamily({
   key: "labelTuples",
   get: (dimension: string) => ({ get }) => {
